corrected the student repo test with the actual tests, moved from the name generator tests
